// 25032026/sce-front-auditoria/sce-front-auditoria/src/app/model/reportes/sistemas-automatizados.ts
export interface FiltroSistemasAutomatizados {
    esquema: string;
    idProceso: number;
    idEleccion: number;    
    idCentroComputo: number;
    idAmbitoElectoral: number;
    ubigeo: string;
    estado: string;
    proceso: string;	
    usuario: string;
    centroComputo: string;    
    odpe: string;
    acronimo: string;
    eleccion: string;
}